app fully working after decomposing Dashboard Component
